+++ CHANGELOG.md
@@ -5,16 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.0] - 2025-07-18
 
 ### Added
 
 - `Rect::intersect`
+- `Add<Pos<T>>` and `AddAssign<Pos<T>>` for `Rect<T>`
 
 ### Changed
 
 - `Layout::iter_pos` and associated methods take ownership of `Rect`
 - `Rect::iter_pos_*` is now `::into_iter_*`, and consumes `Rect`
+- `Rect::from_ltwh` now takes `usize` for `w` and `h` and is infallible
+- `Size` is now always `usize`-based dimensions, removing the `Int` generic
+
+### Removed
+
+- `IntoSize` in favor of `TryFrom<Pos<T>>` for `Size`
 
 ## [0.2.0] - 2025-07-12